refactor(training): migrate state to Pinia
