--- src/store/training/index.js
@@ -1,13 +1,0 @@
-import actions from "./actions";
-import getters from "./getters";
-import mutations from "./mutations";
-import state from "./state";
-
-export default {
-  namespaced: true,
-
-  state,
-  getters,
-  mutations,
-  actions,
-};
