--- src/store/training/state/index.js
@@ -1,7 +1,0 @@
-export default {
-  loading: false,
-  gyms: {},
-  data: {},
-
-  dataRetrieved: false,
-};
