--- src/store/training/types.js
@@ -1,11 +1,0 @@
-export const Action = {
-  GET_DATA: "GET_DATA",
-};
-
-export const Mutation = {
-  ADD_SESSION: "ADD_SESSION",
-  SET_DATA: "SET_DATA",
-  SET_DATA_MISSING: "SET_DATA_MISSING",
-  SET_DATA_RETRIEVED: "SET_DATA_RETRIEVED",
-  SET_GYMS: "SET_GYMS",
-};
